requirements.txt
filelock==3.16.1
fsspec==2024.9.0
Jinja2==3.1.4
MarkupSafe==2.1.5
maturin==1.7.4
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
pillow==10.4.0
setuptools==75.1.0
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
typing_extensions==4.12.2
